feat: Transition to PostgreSQL for data storage, removing JSON fallback and hybrid storage service
