+++ .github/copilot-instructions.md
@@ -9,13 +9,13 @@ Discord bot application that monitors Minecraft server deaths via FTP log parsin
 - **Language**: TypeScript with Node.js 18+
 - **Discord**: discord.js for bot framework and slash commands
 - **Log Access**: FTP client for real-time Minecraft server log monitoring
-- **Storage**: JSON file-based persistence (players.json, config.json, log state tracking)
+- **Storage**: PostgreSQL database for all data persistence (DATABASE_URL required)
 - **Config**: dotenv for environment variable management
 
 ## Architecture Principles
 
 - Single feature focus with planned leaderboard extension
-- Simple file-based data storage
+- Database-only data storage with PostgreSQL
 - Event-driven Discord bot architecture
 - Graceful error handling without crashes
 - 30-second rate limiting per player
@@ -39,9 +39,10 @@ Discord bot application that monitors Minecraft server deaths via FTP log parsin
 - FTP log parsing every 10 seconds for accurate death detection with real causes
 - Rate limit: ignore deaths <30 seconds apart from same player
 - Comprehensive death message parsing with regex patterns
-- Environment variables for all secrets (tokens, passwords, FTP credentials)
+- Environment variables for all secrets (tokens, passwords, FTP credentials, DATABASE_URL)
 - Auto-reconnection logic for FTP connections
-- Log position persistence prevents duplicate announcements on bot restart
+- Database persistence prevents duplicate announcements on bot restart
+- PostgreSQL required for all data storage - no JSON fallback
 
 ## File Structure
 
@@ -50,7 +51,7 @@ src/
 ├── bot.ts          # Main bot entry point and Discord client
 ├── logParser.ts    # FTP log monitoring and death parsing
 ├── discord.ts      # Message formatting and embed creation
-├── storage.ts      # JSON file data persistence with log state tracking
+├── database.ts     # PostgreSQL database service implementing IStorageService
 ├── playerTracker.ts # Player death statistics and rate limiting
 ├── announcer.ts    # Discord announcement service
 ├── leaderboardService.ts # Daily leaderboard generation and scheduling
@@ -72,6 +73,7 @@ src/
 - 2025-09-17: Enhanced timezone handling and rate limiting accuracy
 - 2025-09-17: Removed RCON dependency entirely - now FTP-only architecture
 - 2025-09-17: Planned daily death leaderboard feature with survival champion tracking
+- 2025-09-19: Removed JSON fallback system - now database-only storage architecture
 
 ## Constitutional Requirements
 
@@ -89,4 +91,5 @@ src/
 4. ✅ Error handling and reconnection logic
 5. ✅ Log position tracking and duplicate prevention
 6. ✅ RCON removal - simplified FTP-only architecture
-7. 🔄 Daily death leaderboard with survival champion tracking
+7. ✅ Database-only storage - removed JSON fallback system
+8. 🔄 Daily death leaderboard with survival champion tracking
